Refactor adapter (WIP)
