+++ strelets/docker_config.go
@@ -1,9 +1,5 @@
 package strelets
 
-import (
-	"strconv"
-)
-
 type DockerImageConfig struct {
 	Image               string
 	Tag                 string
@@ -13,56 +9,4 @@ type DockerImageConfig struct {
 
 func (c *DockerImageConfig) FullImageName() string {
 	return c.Image + ":" + c.Tag
-}
-
-func buildDockerNetworkOptions(httpPort int, gossipPort int) (exposedPorts map[string]interface{}, portBindings map[string][]portBinding) {
-	exposedPorts = make(map[string]interface{})
-	exposedPorts["8080/tcp"] = struct{}{}
-	exposedPorts["4400/tcp"] = struct{}{}
-
-	portBindings = make(map[string][]portBinding)
-	portBindings["8080/tcp"] = []portBinding{{"0.0.0.0", strconv.FormatInt(int64(httpPort), 10)}}
-	portBindings["4400/tcp"] = []portBinding{{"0.0.0.0", strconv.FormatInt(int64(gossipPort), 10)}}
-
-	return
-}
-
-func buildDockerConfig(
-	imageName string,
-	exposedPorts map[string]interface{},
-	portBindings map[string][]portBinding,
-	volumes *virtualChainVolumes,
-) map[string]interface{} {
-	configMap := make(map[string]interface{})
-	configMap["Image"] = imageName
-	configMap["ExposedPorts"] = exposedPorts
-	configMap["CMD"] = []string{
-		"/opt/orbs/orbs-node",
-		"--silent",
-		"--config", "/opt/orbs/config/keys.json",
-		"--config", "/opt/orbs/config/network.json",
-		"--log", "/opt/orbs/logs/node.log",
-	}
-
-	hostConfigMap := make(map[string]interface{})
-	hostConfigMap["Binds"] = []string{
-		volumes.keyPairConfigFile + ":/opt/orbs/config/keys.json",
-		volumes.networkConfigFile + ":/opt/orbs/config/network.json",
-		volumes.logsDir + ":/opt/orbs/logs/",
-	}
-	hostConfigMap["PortBindings"] = portBindings
-
-	configMap["HostConfig"] = hostConfigMap
-
-	return configMap
-}
-
-type portBinding struct {
-	HostIp   string
-	HostPort string
-}
-
-func (v *VirtualChain) getContainerConfig(root string) map[string]interface{} {
-	exposedPorts, portBindings := buildDockerNetworkOptions(v.HttpPort, v.GossipPort)
-	return buildDockerConfig(v.DockerConfig.FullImageName(), exposedPorts, portBindings, v.getContainerVolumes(root))
 }
