+++ strelets/virtual_chain.go
@@ -3,8 +3,6 @@ package strelets
 import (
 	"fmt"
 	"io/ioutil"
-	"os"
-	"path/filepath"
 )
 
 type VirtualChainId uint32
@@ -20,10 +18,6 @@ func (v *VirtualChain) getContainerName() string {
 	return fmt.Sprintf("%s-chain-%d", v.DockerConfig.ContainerNamePrefix, v.Id)
 }
 
-func createDir(path string) error {
-	return os.MkdirAll(path, 0755)
-}
-
 func copyFile(source string, destination string) error {
 	data, err := ioutil.ReadFile(source)
 	if err != nil {
@@ -31,34 +25,4 @@ func copyFile(source string, destination string) error {
 	}
 
 	return ioutil.WriteFile(destination, data, 0600)
-}
-
-type virtualChainVolumes struct {
-	configRootDir string
-	logsDir       string
-
-	keyPairConfigFile string
-	networkConfigFile string
-}
-
-func (v *virtualChainVolumes) createDirs() {
-	createDir(v.configRootDir)
-	createDir(v.logsDir)
-}
-
-func (v *VirtualChain) getContainerVolumes(root string) *virtualChainVolumes {
-	containerName := v.getContainerName()
-
-	absolutePathToConfigDir := filepath.Join(root, containerName, "config")
-	absolutePathToLogDir, _ := filepath.Abs(filepath.Join(root, containerName, "logs"))
-
-	absolutePathToNetworkConfig, _ := filepath.Abs(filepath.Join(absolutePathToConfigDir, "network.json"))
-	absolutePathToKeyPairConfig, _ := filepath.Abs(filepath.Join(absolutePathToConfigDir, "keys.json"))
-
-	return &virtualChainVolumes{
-		configRootDir:     absolutePathToConfigDir,
-		logsDir:           absolutePathToLogDir,
-		keyPairConfigFile: absolutePathToKeyPairConfig,
-		networkConfigFile: absolutePathToNetworkConfig,
-	}
 }
