+++ test.e2e.sh
@@ -1,6 +1,7 @@
 #!/bin/bash -x
 
 docker rm -f $(docker ps -aq)
+docker service rm $(docker service ls -q)
 rm -rf _tmp
 
 E2E_CONFIG=./e2e-config/ ./e2e.test -test.v
